+++ ui.cpp
@@ -181,9 +181,9 @@ void rgba_from_rgb(uint32_t v) {
 void draw(game &g) {
 	color_from_rgb(0x666666);
 	SDL_RenderFillRect(renderer,&bg);
-	for (int_fast8_t i = 0; i < 10; i++)
+	for (int i = 0; i < 10; i++)
 	{
-		for (int_fast8_t j = 0; j < 21; j++)
+		for (int j = 0; j < 21; j++)
 		{
 			rect.x =1+ BOARDX+i * (block_size + 1);
 			rect.y = -block_size/2+j * (block_size + 1);
@@ -193,13 +193,13 @@ void draw(game &g) {
 		}
 	}
 	// queue
-	for (int_fast8_t n = 0; n < 5; n++)
+	for (int n = 0; n < 5; n++)
 	{
 
 		color_from_rgb(colors[g.queue[n]+1]);
-		for (int_fast8_t i = 0; i < 4; i++)
+		for (int i = 0; i < 4; i++)
 		{
-			for (int_fast8_t j = 0; j < 4; j++)
+			for (int j = 0; j < 4; j++)
 			{
 				rect.x = 4*BOARDX + i * (block_size + 1);
 				rect.y = block_size*3*n+j * (block_size + 1);
@@ -215,9 +215,9 @@ void draw(game &g) {
 	if (g.held_piece != -1)
 	{
 		color_from_rgb(colors[g.held_piece+1]);
-		for (int_fast8_t i = 0; i < 4; i++)
+		for (int i = 0; i < 4; i++)
 		{
-			for (int_fast8_t j = 0; j < 4; j++)
+			for (int j = 0; j < 4; j++)
 			{
 				if (g.piecedefs[g.held_piece][0][j][i] != -1){
 					
@@ -232,9 +232,9 @@ void draw(game &g) {
 	//active
 	if (active_piece)
 	{
-	for (int_fast8_t i = 0; i < 4; i++)
+	for (int i = 0; i < 4; i++)
 	{
-		for (int_fast8_t j = 0; j < 4; j++)
+		for (int j = 0; j < 4; j++)
 		{
 			rect.x = BOARDX + (g.x + i) * (block_size + 1);
 			rect.y = block_size/2+(g.y - 10 + j) * (block_size + 1);
@@ -249,9 +249,9 @@ void draw(game &g) {
 	}
 	else
 	{
-		for (int_fast8_t i = 0; i < 4; i++)
+		for (int i = 0; i < 4; i++)
 		{
-			for (int_fast8_t j = 0; j < 4; j++)
+			for (int j = 0; j < 4; j++)
 			{
 				rect.x = BOARDX + (3 + i) * (block_size + 1);
 				rect.y = block_size / 2 + (-1 + j) * (block_size + 1);
@@ -268,9 +268,9 @@ void draw(game &g) {
 	//ghost
 	if (ghost){
 		ghosty= g.y+g.softdropdist();
-		for (int_fast8_t i = 0; i < 4; i++)
+		for (int i = 0; i < 4; i++)
 		{
-			for (int_fast8_t j = 0; j < 4; j++)
+			for (int j = 0; j < 4; j++)
 			{
 				rect.x = BOARDX + (g.x + i) * (block_size + 1);
 				rect.y = block_size/2+(ghosty - 10 + j) * (block_size + 1);
